# Feature/barber initial mockup