--- a/assets/css/style.css
+++ b/assets/css/style.css
@@ -1,0 +1,278 @@
+@import url(reset.css);
+
+:root {
+	/* COLORS */
+
+	--rich-black-fogra-39_50: hsla(0, 0%, 5%, 0.5);
+	--rich-black-fogra-39: hsl(0, 0%, 5%);
+	--indian-yellow_10: hsla(36, 61%, 58%, 0.1);
+	--indian-yellow: hsl(36, 61%, 58%);
+	--harvest-gold: hsl(36, 66%, 53%);
+	--eerie-black-1: hsl(0, 0%, 14%);
+	--eerie-black-2: hsl(0, 0%, 12%);
+	--eerie-black-2_85: hsla(0, 0%, 12%, 0.85);
+	--eerie-black-3: hsl(0, 0%, 8%);
+	--sonic-silver: hsl(0, 0%, 44%);
+	--davys-gray: hsl(210, 9%, 31%);
+	--light-gray: hsl(0, 0%, 80%);
+	--platinum: hsl(0, 0%, 91%);
+	--black_30: hsla(0, 0%, 0%, 0.3);
+	--white_10: hsla(0, 0%, 100%, 0.1);
+	--white_30: hsla(0, 0%, 100%, 0.3);
+	--white_50: hsla(0, 0%, 100%, 0.5);
+	--white: hsl(0, 0%, 100%);
+	--jet: hsl(0, 0%, 21%);
+
+	/* TYPOGRAPHY */
+
+	--ff-oswald: "Oswald", sans-serif;
+	--ff-rubik: "Rubik", sans-serif;
+
+	--fs-40: 4rem;
+	--fs-30: 3rem;
+	--fs-24: 2.4rem;
+	--fs-18: 1.8rem;
+	--fs-14: 1.4rem;
+	--fs-13: 1.3rem;
+
+	--fw-300: 300;
+	--fw-500: 500;
+	--fw-600: 600;
+	--fw-700: 700;
+
+	/* SPACING */
+
+	--section-padding: 80px;
+
+	/* SHADOW */
+
+	--shadow-1: 10px 0 60px hsla(0, 0%, 15%, 0.07);
+	--shadow-2: 10px 0 60px hsla(0, 0%, 15%, 0.1);
+
+	/* BORDER RADIUS */
+
+	--radius-5: 5px;
+	--radius-8: 8px;
+
+	/* TRANSITION */
+
+	--transition-1: 0.25s ease;
+	--transition-2: 0.5s ease;
+	--cubic-out: cubic-bezier(0.33, 0.85, 0.4, 0.96);
+}
+
+
+/*-----------------------------------*\
+  #REUSED STYLE
+\*-----------------------------------*/
+
+.container { padding-inline: 15px; }
+
+.section { padding-block: var(--section-padding); }
+
+.has-before,
+.has-after {
+  position: relative;
+  z-index: 1;
+}
+
+.has-before::before,
+.has-after::after {
+  position: absolute;
+  content: "";
+}
+
+.has-bg-image {
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-position: center;
+}
+
+.h1,
+.h2,
+.h3 {
+  font-family: var(--ff-oswald);
+  line-height: 1.3;
+}
+
+.h1,
+.h2 { text-transform: uppercase; }
+
+.h1,
+.h3 { font-weight: var(--fw-600); }
+
+.h1 {
+  color: var(--white);
+  font-size: var(--fs-40);
+}
+
+.h2,
+.h3 { color: var(--eerie-black-1); }
+
+.h2 { font-size: var(--fs-30); }
+
+.h3 { font-size: var(--fs-24); }
+
+.btn {
+  color: var(--white);
+  background-color: var(--indian-yellow);
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  max-width: max-content;
+  padding: 10px 25px;
+  font-family: var(--ff-oswald);
+  font-size: var(--fs-14);
+  font-weight: var(--fw-600);
+  text-transform: uppercase;
+  border-radius: var(--radius-5);
+  overflow: hidden;
+}
+
+.btn::before {
+  background-color: var(--eerie-black-1);
+  inset: 0;
+  z-index: -1;
+  transform: skewY(-15deg) scaleY(0);
+  transition: var(--transition-2);
+}
+
+.btn:is(:hover, :focus)::before { transform: skewY(-15deg) scaleY(2.5); }
+
+.text-center { text-align: center; }
+
+.grid-list {
+  display: grid;
+  gap: 30px;
+}
+
+.img-holder {
+  aspect-ratio: var(--width) / var(--height);
+  background-color: var(--light-gray);
+  overflow: hidden;
+}
+
+.img-cover {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+
+
+
+
+/*-----------------------------------*\
+  #HEADER
+\*-----------------------------------*/
+
+.header-top-item,
+.header .btn { display: none; }
+
+.header-top-list,
+.header-top-list .social-list {
+  display: flex;
+  align-items: center;
+}
+
+.header-top-list { justify-content: center; }
+
+.header-top-list .social-list {
+  gap: 20px;
+  padding-block: 15px;
+}
+
+.header-top-list .social-link {
+  color: var(--sonic-silver);
+  font-size: 15px;
+  transition: var(--transition-1);
+}
+
+.header-top-list .social-link:is(:hover, :focus) { color: var(--indian-yellow); }
+
+.header-bottom {
+  position: absolute;
+  top: 45px;
+  left: 0;
+  width: 100%;
+  padding-block: 10px;
+  z-index: 4;
+}
+
+.header-bottom.active {
+  position: fixed;
+  top: 0;
+  background-color: var(--rich-black-fogra-39);
+  transform: translateY(-100%);
+  animation: slideIn 0.5s ease forwards;
+}
+
+@keyframes slideIn {
+  0% { transform: translateY(-100%); }
+  100% { transform: translateY(0); }
+}
+
+.header-bottom > .container {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.logo {
+  color: var(--indian-yellow);
+  font-family: var(--ff-oswald);
+  font-size: 2.7rem;
+  font-weight: var(--fw-600);
+  text-transform: uppercase;
+  line-height: 1.4;
+}
+
+.logo .span {
+  color: var(--white);
+  font-family: var(--ff-rubik);
+  font-size: 1.3rem;
+  font-weight: var(--fw-300);
+  text-transform: capitalize;
+  letter-spacing: 2px;
+}
+
+.nav-toggle-btn {
+  color: var(--white);
+  font-size: 40px;
+}
+
+.navbar {
+  position: absolute;
+  padding-inline: 0;
+  top: 100%;
+  left: 15px;
+  right: 15px;
+  background-color: var(--rich-black-fogra-39);
+  max-height: 0;
+  overflow: hidden;
+  transition: 0.15s var(--cubic-out);
+}
+
+.navbar.active {
+  max-height: 321px;
+  transition-duration: 0.5s;
+}
+
+.navbar-list {
+  border-block-start: 1px solid var(--jet);
+  margin-block: 25px;
+}
+
+.navbar-item { border-block-end: 1px solid var(--jet); }
+
+.navbar-link {
+  color: var(--white);
+  font-family: var(--ff-oswald);
+  font-weight: var(--fw-600);
+  text-transform: uppercase;
+  line-height: 1.5;
+  padding: 10px 30px;
+  transition: var(--transition-1);
+}
+
+.navbar-link:is(:hover, :focus) { color: var(--indian-yellow); }