--- a/assets/js/core/Observable.js
+++ b/assets/js/core/Observable.js
@@ -1,0 +1,19 @@
+class Observable {
+    constructor() {
+        this.observers = [];
+    }
+
+    subscribe(observer) {
+        this.observers.push(observer);
+    }
+
+    unsubscribe(observer) {
+        this.observers = this.observers.filter(obs => obs !== observer);
+    }
+
+    notify(data) {
+        for (const observer of this.observers) observer.update(data);
+    }
+}
+
+export default Observable;