--- a/index.html
+++ b/index.html
@@ -1,0 +1,179 @@
+<!DOCTYPE html>
+<html lang="es-ES" dir="ltr" itemscope itemtype="https://schema.org/BarberShop">
+
+<head>
+    <!-- Metadatos esenciales -->
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
+    <meta name="theme-color" content="#1a1a1a">
+    <meta http-equiv="x-ua-compatible" content="ie=edge">
+
+    <!-- OpenGraph optimizado -->
+    <meta property="og:site_name" content="Deluxe Barber">
+    <meta property="og:title" content="Deluxe Barber | Barbería Premium en [Ciudad] | Cortes Expertos y Diseño de Barba">
+    <meta property="og:description" content="✂️ Barbería de lujo en [Ciudad] | Reserva online tu cita | Cortes de pelo personalizados, arreglo de barba profesional y tratamientos faciales premium | Profesionales certificados 🪒">
+    <meta property="og:type" content="website">
+    <meta property="og:url" content="https://www.tudominio.com/barberia-premium">
+    <meta property="og:image" content="https://www.tudominio.com/images/og-deluxe-barber-cortes-premium.jpg">
+    <meta property="og:image:width" content="1200">
+    <meta property="og:image:height" content="630">
+    <meta property="og:image:alt" content="Barbero profesional realizando un corte de pelo estilo fade en Deluxe Barber [Ciudad]">
+    <meta property="og:locale" content="es_ES">
+    <meta property="og:locale:alternate" content="es_MX">
+    <meta property="og:locale:alternate" content="es_AR">
+
+    <!-- Twitter Card optimizado -->
+    <meta name="twitter:card" content="summary_large_image">
+    <meta name="twitter:site" content="@DeluxeBarber_[Ciudad]">
+    <meta name="twitter:creator" content="@JefeBarbero_[Ciudad]">
+    <meta name="twitter:title" content="Cortes Premium en [Ciudad] | Deluxe Barber | Reserva Online 24/7">
+    <meta name="twitter:description" content="🪒 Tu estilo merece lo mejor! Barbería premium con técnicas tradicionales y modernas | Reserva instantánea | Abiertos de [horario] 💈">
+    <meta name="twitter:image" content="https://www.tudominio.com/images/twitter-card-deluxe-barber.jpg">
+    <meta name="twitter:image:alt" content="Salón de barbería Deluxe Barber con equipos profesionales y ambiente exclusivo en [Ciudad]">
+
+    <!-- SEO Avanzado -->
+    <meta name="description" content="Barbería premium en [Ciudad]. Especialistas en cortes modernos, diseño de barba y tratamientos faciales. Reserva online con descuento en tu primera cita ✂️ Profesionales certificados 🪒 Atención personalizada">
+    <link href="https://www.tudominio.com/barberia-premium" rel="canonical">
+    <meta name="robots" content="index, follow, max-image-preview:large">
+    <meta name="revisit-after" content="7 days">
+    <meta name="geo.region" content="ES-M">
+    <meta name="geo.placename" content="[Ciudad]">
+    <meta name="author" content="Deluxe Barber Team">
+    <meta name="rating" content="safe for kids">
+
+
+    <!-- Favicon y recursos -->
+    <link href="favicon.svg" rel="icon" sizes="48x48">
+    <link type="image/svg+xml" href="/icon.svg" rel="icon" sizes="any">
+    <link href="/apple-touch-icon.png" rel="apple-touch-icon">
+    <link href="/site.webmanifest" rel="manifest">
+    <link href="assets/css/flaticon.min.css" rel="stylesheet">
+    <link href="assets/css/style.css" rel="stylesheet">
+
+    <!-- Preconexiones críticas -->
+    <link href="https://fonts.googleapis.com" rel="preconnect">
+    <link href="https://fonts.gstatic.com" rel="preconnect" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Oswald:wght@400;500;600;700&family=Rubik:wght@300,400;700&display=swap" rel="stylesheet">
+    <link href="https://cdn.tudominio.com" rel="preconnect" crossorigin>
+    <script src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.esm.js" type="module"></script>
+    <script src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.js" nomodule></script>
+
+    <title>Deluxe Barber</title>
+</head>
+
+<body>
+
+    <header class="header">
+
+        <div class="header-top">
+          <div class="container">
+    
+            <ul class="header-top-list">
+    
+              <li class="header-top-item">
+                <ion-icon name="call-outline" aria-hidden="true"></ion-icon>
+    
+                <p class="item-title">Call Us :</p>
+    
+                <a href="tel:01234567895" class="item-link">012 (345) 67 895</a>
+              </li>
+    
+              <li class="header-top-item">
+                <ion-icon name="time-outline" aria-hidden="true"></ion-icon>
+    
+                <p class="item-title">Opening Hour :</p>
+    
+                <p class="item-text">Sunday - Friday, 08 am - 09 pm</p>
+              </li>
+    
+              <li>
+                <ul class="social-list">
+    
+                  <li>
+                    <a href="#" class="social-link">
+                      <ion-icon name="logo-facebook"></ion-icon>
+                    </a>
+                  </li>
+    
+                  <li>
+                    <a href="#" class="social-link">
+                      <ion-icon name="logo-twitter"></ion-icon>
+                    </a>
+                  </li>
+    
+                  <li>
+                    <a href="#" class="social-link">
+                      <ion-icon name="logo-youtube"></ion-icon>
+                    </a>
+                  </li>
+    
+                  <li>
+                    <a href="#" class="social-link">
+                      <ion-icon name="chatbubble-ellipses-outline"></ion-icon>
+                    </a>
+                  </li>
+    
+                </ul>
+              </li>
+    
+            </ul>
+    
+          </div>
+        </div>
+    
+        <div class="header-bottom" data-header>
+          <div class="container">
+    
+            <a href="#" class="logo">
+              Barber
+              <span class="span">Hair Salon</span>
+            </a>
+    
+            <nav class="navbar container" data-navbar>
+              <ul class="navbar-list">
+    
+                <li class="navbar-item">
+                  <a href="#home" class="navbar-link" data-nav-link>Home</a>
+                </li>
+    
+                <li class="navbar-item">
+                  <a href="#services" class="navbar-link" data-nav-link>Services</a>
+                </li>
+    
+                <li class="navbar-item">
+                  <a href="#pricing" class="navbar-link" data-nav-link>Pricing</a>
+                </li>
+    
+                <li class="navbar-item">
+                  <a href="#gallery" class="navbar-link" data-nav-link>Gallery</a>
+                </li>
+    
+                <li class="navbar-item">
+                  <a href="#appointment" class="navbar-link" data-nav-link>Appointment</a>
+                </li>
+    
+                <li class="navbar-item">
+                  <a href="#" class="navbar-link" data-nav-link>Contact</a>
+                </li>
+    
+              </ul>
+            </nav>
+    
+            <button class="nav-toggle-btn" aria-label="toggle menu" data-nav-toggler>
+              <ion-icon name="menu-outline" aria-hidden="true"></ion-icon>
+            </button>
+    
+            <a href="#" class="btn has-before">
+              <span class="span">Appointment</span>
+    
+              <ion-icon name="arrow-forward" aria-hidden="true"></ion-icon>
+            </a>
+    
+          </div>
+        </div>
+    
+      </header>
+
+</body>
+
+</html>