… assets/css/flaticon.min.css
… assets/fonts/flaticon.ttf
… assets/images/appoin-banner-1.jpg
… assets/images/appoin-banner-2.jpg
… assets/images/footer-bg.png
… assets/images/footer-blog-1.jpg
… assets/images/footer-blog-2.jpg
… assets/images/gallery-1.jpg
… assets/images/gallery-2.jpg
… assets/images/gallery-3.jpg
… assets/images/gallery-4.jpg
… assets/images/hero-banner.jpg
… assets/images/pricing-1.jpg
… assets/images/pricing-2.jpg
… assets/images/pricing-3.jpg
… assets/images/pricing-4.jpg
… assets/images/pricing-5.jpg
… assets/images/pricing-6.jpg
… assets/images/pricing-7.jpg
… assets/images/pricing-8.jpg
… assets/images/pricing-bg.jpg